+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+xxxxxxxxxxxxxxxxxxx
+===================
+- Added `dataset.map()`, `dataset.forEach()` and `dataset.reduce()` functions.
+
 0.5.34 / 2018-05-18
 ===================
 - Added `loggingIntervalMillis` options to `AutoscaledPool`